requirements.txt
azure-iot-device
azure-eventhub
azure-storage-queue==12.1.5
azureml-core
azureml-dataprep
boto3~=1.16.49
botocore~=1.19.49
click==7.1.2
environs
pydub~=0.24.1
pytest
pandas==0.25.3
matplotlib~=3.3.3
scipy==1.4.1
numpy~=1.19.2
librosa==0.8.0
noisereduce
soundfile==0.10.3.post1
split-folders==0.4.3
opencv-python==4.5.1.48
flask==1.1.2
flask-googlemaps==0.4.1
flask-script==2.0.6
flask-sqlalchemy==2.4.4
flask-migrate==2.7.0
tensorflow==2.2.0
